Fixed userCollectionPath errors, activity vs fragment tagselection
