+++ app/javaDocs/com/example/cmput301groupproject/ItemFragment.html
@@ -73,7 +73,7 @@ loadScripts(document, 'script');</script>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">androidx.fragment.app.Fragment
 <div class="inheritance">androidx.fragment.app.DialogFragment
-<div class="inheritance">com.example.cmput301groupproject.ItemEditFragment</div>
+<div class="inheritance">com.example.cmput301groupproject.ItemEditActivity</div>
 </div>
 </div>
 </div>
